# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/scan": {"post": {"summary": "Scan", "operationId": "scan_api_scan_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScanRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScanResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/scans": {"get": {"summary": "Get Scans", "operationId": "get_scans_api_scans_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListScansResponse"}}}}}}}, "/api/scans/{scan_id}": {"get": {"summary": "Get Scan Details", "operationId": "get_scan_details_api_scans__scan_id__get", "parameters": [{"name": "scan_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Scan Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetScanResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/generate-report": {"post": {"summary": "Generate Report", "operationId": "generate_report_api_generate_report_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateReportRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateReportResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/buckets": {"get": {"summary": "List Buckets", "operationId": "list_buckets_api_buckets_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BucketsResponse"}}}}}}}, "/api/files": {"get": {"summary": "List Files", "operationId": "list_files_api_files_get", "parameters": [{"name": "bucket", "in": "query", "required": false, "schema": {"type": "string", "title": "Bucket"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FilesResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/validate": {"post": {"summary": "Validate Credentials", "operationId": "validate_credentials_api_validate_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ValidateRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ValidateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"BucketsResponse": {"properties": {"buckets": {"items": {"type": "string"}, "type": "array", "title": "Buckets"}}, "type": "object", "required": ["buckets"], "title": "BucketsResponse"}, "FilesResponse": {"properties": {"files": {"items": {"type": "string"}, "type": "array", "title": "Files"}}, "type": "object", "required": ["files"], "title": "FilesResponse"}, "GenerateReportRequest": {"properties": {"scan_id": {"type": "string", "title": "Scan Id"}, "format": {"$ref": "#/components/schemas/ReportFormat", "default": "pdf"}}, "type": "object", "required": ["scan_id"], "title": "GenerateReportRequest"}, "GenerateReportResponse": {"properties": {"report_url": {"type": "string", "title": "Report Url"}}, "type": "object", "required": ["report_url"], "title": "GenerateReportResponse"}, "GetScanResponse": {"properties": {"findings": {"items": {"$ref": "#/components/schemas/VulnerabilityFinding"}, "type": "array", "title": "Findings"}}, "type": "object", "title": "GetScanResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ListScansResponse": {"properties": {"scans": {"items": {"$ref": "#/components/schemas/ScanItem"}, "type": "array", "title": "Scans"}}, "type": "object", "required": ["scans"], "title": "ListScansResponse"}, "ReportFormat": {"type": "string", "enum": ["pdf", "json", "csv"], "title": "ReportFormat"}, "ScanItem": {"properties": {"scan_id": {"type": "string", "title": "Scan Id"}, "access_key": {"type": "string", "title": "Access Key"}, "selected_scans": {"items": {"items": {"type": "string"}, "type": "array"}, "type": "array", "title": "Selected Scans"}, "findings": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Findings"}, "completed_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Completed At"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["scan_id", "access_key", "selected_scans", "created_at"], "title": "ScanItem"}, "ScanRequest": {"properties": {"bucket": {"type": "string", "title": "Bucket"}, "file": {"type": "string", "title": "File"}, "services": {"additionalProperties": {"items": {"type": "string"}, "type": "array"}, "type": "object", "title": "Services", "default": {}}}, "type": "object", "title": "ScanRequest"}, "ScanResponse": {"properties": {"scan_id": {"type": "string", "title": "Scan Id"}, "findings": {"items": {"$ref": "#/components/schemas/VulnerabilityFinding"}, "type": "array", "title": "Findings"}}, "type": "object", "required": ["scan_id"], "title": "ScanResponse"}, "ValidateRequest": {"properties": {"access_key": {"type": "string", "title": "Access Key"}, "secret_key": {"type": "string", "title": "Secret Key"}, "region": {"type": "string", "title": "Region"}}, "type": "object", "required": ["access_key", "secret_key", "region"], "title": "ValidateRequest"}, "ValidateResponse": {"properties": {"valid": {"type": "boolean", "title": "Valid"}}, "type": "object", "required": ["valid"], "title": "ValidateResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "Vulnerability": {"type": "string", "enum": ["public_s3_bucket", "unencrypted_s3_bucket", "s3_bucket_versioning_disabled", "s3_bucket_logging_disabled", "s3_bucket_block_public_access_disabled", "over_permissive_iam", "iam_user_no_mfa", "iam_unused_access_key", "iam_inline_policy", "iam_root_access_key", "open_security_group_ingress", "open_security_group_egress", "unused_security_group", "cloudtrail_not_logging", "cloudtrail_not_multi_region", "cloudtrail_no_log_file_validation", "cloudtrail_bucket_public", "cloudtrail_bucket_encryption_disabled", "guardduty_disabled", "vpc_flow_logs_disabled", "ebs_volume_unencrypted", "rds_instance_unencrypted", "rds_instance_public_access", "ssm_parameter_unencrypted", "lambda_overpermissive_role", "lambda_public_access", "apigateway_open_resource", "iam_user_with_console_access", "ec2_instance_public_ip"], "title": "Vulnerability"}, "VulnerabilityFinding": {"properties": {"type": {"$ref": "#/components/schemas/Vulnerability"}, "name": {"type": "string", "title": "Name"}, "severity": {"type": "string", "title": "Severity"}, "details": {"type": "string", "title": "Details"}}, "type": "object", "required": ["type", "name", "severity", "details"], "title": "VulnerabilityFinding"}}}}